fix(#778): drop VST3 plugins on the main thread so an open native editor can't crash
With a VST3 block's native editor open and the chain running, changing a
parameter SIGSEGV'd: the superseded runtime (holding the `Vst3Plugin` whose
JUCE editor is open) is dropped on the `openrig-control-worker` thread (#672),
and `Vst3Plugin::drop` -> `terminate()` tears down the plugin's editor
components (`~SliderLabelComp` -> `makeKeyWindow`) off the main thread, which
macOS AppKit forbids. Confirmed by the crash backtrace.

- `vst3-host`: new `main_thread` module — `mark_main_thread()` + a deferred
  teardown queue. `Vst3Plugin` now wraps its COM state in `ManuallyDrop<Vst3Inner>`;
  `Drop` runs the teardown inline on the main thread (or when no GUI is
  registered — CLI/tests/render), else hands it to the queue.
- `adapter-gui`: mark the UI thread at startup and drain the queue on the
  frontend tick (right after `poll_pending_rebuilds`).
- Split `Vst3Plugin::load` into `host_load.rs` to keep `host.rs` under the
  per-file line cap.

The crash only reproduces through the full app editor flow (five headless
harnesses could not isolate it), so the guard here is the deferral mechanism
(`main_thread_tests.rs`); final confirmation is the app: add ChowCentaur, run
the chain, open its editor, move a knob -> no crash.
